requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.2
numpy==1.21.2
pandas==1.3.3
pkg-resources==0.0.0
PyJWT==2.2.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
telebot==0.0.4
urllib3==1.26.7
